Remove separate kotlinX source set for extensions
