+++ library/src/main/java/com/getstream/sdk/chat/ChatSocketListener.kt
@@ -1,6 +1,7 @@
 package com.getstream.sdk.chat
 
 import com.getstream.sdk.chat.enums.OnlineStatus
+import com.getstream.sdk.chat.utils.exhaustive
 import io.getstream.chat.android.client.errors.ChatError
 import io.getstream.chat.android.client.events.ChannelCreatedEvent
 import io.getstream.chat.android.client.events.ChannelDeletedEvent
