+++ library/src/main/java/com/getstream/sdk/chat/utils/Date.kt
@@ -1,3 +1,5 @@
-package java.util
+package com.getstream.sdk.chat.utils
+
+import java.util.Date
 
 internal fun Date.isInLastMinute(): Boolean = (Date().time - 60000 < time)
