+++ library/src/main/java/com/getstream/sdk/chat/utils/Primitives.kt
@@ -1,7 +1,7 @@
-package com.getstream.sdk.chat
+package com.getstream.sdk.chat.utils
 
 internal inline fun Boolean.whenTrue(crossinline f: () -> Unit): Boolean = also { if (this) f() }
 internal inline fun Boolean.whenFalse(crossinline f: () -> Unit): Boolean = also { if (!this) f() }
 
-internal val <T> T.exhaustive: T
+internal inline val <T> T.exhaustive: T
     get() = this
